polish up ux
