+++ src/components/RightActionSidebar.tsx
@@ -1,36 +1,7 @@
 import { useAtom, useAtomValue } from "jotai";
 import { previewModeAtom, selectedAppIdAtom } from "../atoms/appAtoms";
-import { ipc } from "@/ipc/types";
-
-import {
-  Eye,
-  Code,
-  MoreVertical,
-  Cog,
-  Trash2,
-  AlertTriangle,
-  Wrench,
-  Globe,
-  Shield,
-} from "lucide-react";
-import { ChatActivityButton } from "@/components/chat/ChatActivity";
-import { useCallback } from "react";
-
-import { useRunApp } from "@/hooks/useRunApp";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  Tooltip,
-  TooltipTrigger,
-  TooltipContent,
-  TooltipProvider,
-} from "@/components/ui/tooltip";
-import { showError, showSuccess } from "@/lib/toast";
-import { useMutation } from "@tanstack/react-query";
+import { Eye, Code, AlertTriangle, Wrench, Globe, Shield } from "lucide-react";
+import { motion } from "framer-motion";
 import { useCheckProblems } from "@/hooks/useCheckProblems";
 import { isPreviewOpenAtom } from "@/atoms/viewAtoms";
 import { useTranslation } from "react-i18next";
@@ -43,7 +14,6 @@ export const RightActionSidebar = () => {
   const [isPreviewOpen, setIsPreviewOpen] = useAtom(isPreviewOpenAtom);
   const selectedAppId = useAtomValue(selectedAppIdAtom);
   const { problemReport } = useCheckProblems(selectedAppId);
-  const { restartApp, refreshAppIframe } = useRunApp();
 
   const selectPanel = (panel: PreviewMode) => {
     if (previewMode === panel) {
@@ -53,31 +23,6 @@ export const RightActionSidebar = () => {
       setIsPreviewOpen(true);
     }
   };
-
-  const onCleanRestart = useCallback(() => {
-    restartApp({ removeNodeModules: true });
-  }, [restartApp]);
-
-  const useClearSessionData = () => {
-    return useMutation({
-      mutationFn: () => {
-        return ipc.system.clearSessionData();
-      },
-      onSuccess: async () => {
-        await refreshAppIframe();
-        showSuccess("Preview data cleared");
-      },
-      onError: (error) => {
-        showError(`Error clearing preview data: ${error}`);
-      },
-    });
-  };
-
-  const { mutate: clearSessionData } = useClearSessionData();
-
-  const onClearSessionData = useCallback(() => {
-    clearSessionData();
-  }, [clearSessionData]);
 
   // Get the problem count for the selected app
   const problemCount = problemReport ? problemReport.problems.length : 0;
@@ -91,7 +36,7 @@ export const RightActionSidebar = () => {
 
   const displayCount = formatProblemCount(problemCount);
 
-  const iconSize = 20;
+  const iconSize = 18;
 
   const renderButton = (
     mode: PreviewMode,
@@ -102,121 +47,79 @@ export const RightActionSidebar = () => {
   ) => {
     const isActive = previewMode === mode && isPreviewOpen;
     return (
-      <Tooltip>
-        <TooltipTrigger
-          render={
-            <button
-              data-testid={testId}
-              className={`no-app-region-drag cursor-pointer relative flex flex-col items-center gap-1 w-14 h-14 justify-center rounded-2xl text-xs font-medium transition-colors ${
-                isActive
-                  ? "bg-sidebar-accent text-sidebar-accent-foreground"
-                  : "hover:bg-sidebar-accent hover:text-sidebar-accent-foreground"
-              }`}
-              onClick={() => selectPanel(mode)}
-            />
-          }
-        >
-          <div className="relative">
-            {icon}
-            {badge}
-          </div>
-          <span className="text-xs">{text}</span>
-        </TooltipTrigger>
-        <TooltipContent side="left">{text}</TooltipContent>
-      </Tooltip>
+      <button
+        data-testid={testId}
+        className={`no-app-region-drag cursor-pointer relative flex flex-col items-center justify-center w-12 h-12 rounded-lg font-medium transition-colors duration-150 active:scale-90 ${
+          isActive
+            ? "text-sidebar-accent-foreground"
+            : "hover:bg-sidebar-accent hover:text-sidebar-accent-foreground"
+        }`}
+        onClick={() => selectPanel(mode)}
+      >
+        {isActive && (
+          <motion.div
+            layoutId="active-sidebar-indicator"
+            className="absolute inset-0 rounded-lg bg-sidebar-accent"
+            transition={{ type: "spring", stiffness: 500, damping: 35 }}
+          />
+        )}
+        <div className="relative z-10">
+          {icon}
+          {badge}
+        </div>
+        <span className="relative z-10 text-[10px] leading-tight mt-0.5 truncate max-w-full">
+          {text}
+        </span>
+      </button>
     );
   };
 
   return (
-    <TooltipProvider delay={0}>
-      <div className="flex flex-col h-full w-[4.5rem] bg-sidebar border-l border-sidebar-border">
-        {/* Main action buttons */}
-        <div className="flex flex-col items-center gap-1 pt-2 flex-1">
-          {renderButton(
-            "preview",
-            <Eye size={iconSize} />,
-            t("preview.title"),
-            "preview-mode-button",
-          )}
-          {renderButton(
-            "problems",
-            <AlertTriangle size={iconSize} />,
-            t("preview.problems"),
-            "problems-mode-button",
-            displayCount && (
-              <span className="absolute -top-1 -right-1 px-1 py-0.5 text-[10px] font-medium bg-red-100 dark:bg-red-900/30 text-red-700 dark:text-red-300 rounded-full min-w-[16px] text-center">
-                {displayCount}
-              </span>
-            ),
-          )}
-          {renderButton(
-            "code",
-            <Code size={iconSize} />,
-            t("preview.code"),
-            "code-mode-button",
-          )}
-          {renderButton(
-            "configure",
-            <Wrench size={iconSize} />,
-            t("preview.configure"),
-            "configure-mode-button",
-          )}
-          {renderButton(
-            "security",
-            <Shield size={iconSize} />,
-            t("preview.security"),
-            "security-mode-button",
-          )}
-          {renderButton(
-            "publish",
-            <Globe size={iconSize} />,
-            t("preview.publish"),
-            "publish-mode-button",
-          )}
-        </div>
-
-        {/* Bottom section with chat activity and more options */}
-        <div className="flex flex-col items-center gap-2 pb-4">
-          <ChatActivityButton />
-          <DropdownMenu>
-            <Tooltip>
-              <TooltipTrigger
-                render={
-                  <DropdownMenuTrigger
-                    data-testid="preview-more-options-button"
-                    className="no-app-region-drag flex items-center justify-center w-10 h-10 rounded-md text-sm hover:bg-sidebar-accent transition-colors"
-                  />
-                }
-              >
-                <MoreVertical size={20} />
-              </TooltipTrigger>
-              <TooltipContent side="left">
-                {t("preview.moreOptions")}
-              </TooltipContent>
-            </Tooltip>
-            <DropdownMenuContent align="end" side="left" className="w-60">
-              <DropdownMenuItem onClick={onCleanRestart}>
-                <Cog size={16} />
-                <div className="flex flex-col">
-                  <span>{t("preview.rebuild")}</span>
-                  <span className="text-xs text-muted-foreground">
-                    {t("preview.rebuildDescription")}
-                  </span>
-                </div>
-              </DropdownMenuItem>
-              <DropdownMenuItem onClick={onClearSessionData}>
-                <Trash2 size={16} />
-                <div className="flex flex-col">
-                  <span>{t("preview.clearCache")}</span>
-                  <span className="text-xs text-muted-foreground">
-                    {t("preview.clearCacheDescription")}
-                  </span>
-                </div>
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
+    <div className="flex flex-col h-full w-16 pl-1 -mr-1.5 bg-sidebar border-l border-sidebar-border">
+      {/* Main action buttons */}
+      <div className="flex flex-col items-center gap-1 pt-2 flex-1">
+        {renderButton(
+          "preview",
+          <Eye size={iconSize} />,
+          t("preview.title"),
+          "preview-mode-button",
+        )}
+        {renderButton(
+          "problems",
+          <AlertTriangle size={iconSize} />,
+          t("preview.problems"),
+          "problems-mode-button",
+          displayCount && (
+            <span className="absolute -top-1 -right-1 px-1 py-0.5 text-[10px] font-medium bg-red-100 dark:bg-red-900/30 text-red-700 dark:text-red-300 rounded-full min-w-[16px] text-center">
+              {displayCount}
+            </span>
+          ),
+        )}
+        {renderButton(
+          "code",
+          <Code size={iconSize} />,
+          t("preview.code"),
+          "code-mode-button",
+        )}
+        {renderButton(
+          "configure",
+          <Wrench size={iconSize} />,
+          t("preview.configure"),
+          "configure-mode-button",
+        )}
+        {renderButton(
+          "security",
+          <Shield size={iconSize} />,
+          t("preview.security"),
+          "security-mode-button",
+        )}
+        {renderButton(
+          "publish",
+          <Globe size={iconSize} />,
+          t("preview.publish"),
+          "publish-mode-button",
+        )}
       </div>
-    </TooltipProvider>
+    </div>
   );
 };
